Fix most logging stuff
