+++ server/commands/teamserver.go
@@ -115,7 +115,7 @@ func closeCmd(serv *server.Server) func(cmd *cobra.Command, args []string) {
 					if err != nil {
 						fmt.Fprintln(cmd.ErrOrStderr(), command.Warn, err)
 					} else {
-						fmt.Fprintf(cmd.OutOrStdout(), command.Info+"Closed %s listener (%s) [%s]", ln.Name, formatSmallID(ln.ID), ln.Description)
+						fmt.Fprintf(cmd.OutOrStdout(), command.Info+"Closed %s listener (%s) [%s]\n", ln.Name, formatSmallID(ln.ID), ln.Description)
 					}
 				}
 			}
